Use stub modules to avoid CPP imports
