+++ src/Grammar.lhs
@@ -20,20 +20,9 @@ Here is our mid-section datatype
 
 > import GenUtils
 > import AbsSyn
-> import ParseMonad
+> import ParseMonad.Class
 > import AttrGrammar
-
-We use the bootstrapped version if it is available
-
-#ifdef HAPPY_BOOTSTRAP
-
-> import AttrGrammarParser.Bootstrapped
-
-#else
-
-> import AttrGrammarParser.Oracle
-
-#endif
+> import AttrGrammarParser
 
 > import ParamRules
