+++ src/Main.lhs
@@ -11,22 +11,11 @@ Path settings auto-generated by Cabal:
 
 > import Paths_happy
 
-> import ParseMonad
+> import ParseMonad.Class
 > import AbsSyn
 > import Grammar
 > import PrettyGrammar
-
-We use the bootstrapped version if it is available
-
-#ifdef HAPPY_BOOTSTRAP
-
-> import Parser.Bootstrapped
-
-#else
-
-> import Parser.Oracle
-
-#endif
+> import Parser
 
 > import First
 > import LALR
