+++ src/ParseMonad.hs
@@ -1,13 +1,8 @@
-module ParseMonad where
+module ParseMonad (module X) where
 
-type Pfunc a = String -> Int -> ParseResult a
-
-class HasLexer token where
-  lexToken :: (token -> Pfunc r) -> Pfunc r
-
-type ParseResult = Either String
-
-class Monad p => ParseMonad p where
-  failP :: (Int -> String) -> p a
-  lineP :: p Int
-  runFromStartP :: p a -> String -> Int -> ParseResult a
+-- We use the bootstrapped version if it is available
+#ifdef HAPPY_BOOTSTRAP
+import ParseMonad.Bootstrapped as X
+#else
+import ParseMonad.Oracle as X
+#endif
