+++ src/Parser/Bootstrapped.ly
@@ -9,7 +9,7 @@ The parser.
 > {
 > {-# OPTIONS_GHC -w #-}
 > module Parser.Bootstrapped (ourParser,AbsSyn) where
-> import ParseMonad
+> import ParseMonad.Class
 > import ParseMonad.Bootstrapped
 > import AbsSyn
 > import Lexer
